+++ processor/src/test/resources/expected/com/example/route/RouterFu_RouteWithParameters.java
@@ -149,7 +149,8 @@ public class RouterFu_RouteWithParameters extends RouteWithParameters implements
 
   @Nonnull
   @Override
-  public String buildRegionEventLocation(@Nonnull final String regionCode, @Nonnull final String eventCode) {
+  public String buildRegionEventLocation(@Nonnull final String regionCode,
+      @Nonnull final String eventCode) {
     final Map<Parameter, String> $fu$_route_params = new HashMap<>();
     $fu$_route_params.put( $fu$_regionCode, regionCode );
     $fu$_route_params.put( $fu$_eventCode, eventCode );